backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
django-filter>=23.3
channels>=4.0.0
channels-redis>=4.1.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.9
django-dbbackup>=4.1.0
